dcp-967 managed access - filter metadata resources by user's allowed projects
